Moved contentful helpers
