+++ cypress/Contentful/Fields/ImageField.js
