+++ cypress/Contentful/Models/PromoModel.js
