+++ cypress/Contentful/Models/SeriesModel.js
