+++ cypress/integration/external_repo_media/series_page_current_series_spec.js
@@ -1,5 +1,5 @@
-import { ContentfulApi } from '../../support/Contentful/ContentfulApi';
-import { ContentfulElementValidator as Element } from '../../support/Cypress/ContentfulElementValidator';
+import { ContentfulApi } from '../../Contentful/ContentfulApi';
+import { ContentfulElementValidator as Element } from '../../Contentful/ContentfulElementValidator';
 
 describe('Tesing the Current Series on the Media/Series page:', function(){
     let currentSeries;
