+++ cypress/integration/locations/locations_current_series_spec.js
@@ -1,4 +1,4 @@
-import { ContentfulApi } from '../../support/Contentful/ContentfulApi';
+import { ContentfulApi } from '../../Contentful/ContentfulApi';
 
 describe('Testing the Current Series on a random Locations page:', function () {
     let currentSeries;
